requirements.txt
txtflask
flask-sqlalchemy
flask-cors
python-dotenv
pymysql
cryptography
werkzeug